requirements.txt
Django==1.11.20
djangorestframework==3.9.4
altgraph==0.16.1
dis3==0.1.2
docx==0.2.4
future==0.17.1
lxml==4.2.5
macholib==1.11
numpy==1.15.1
pefile==2018.8.8
Pillow==5.2.0
psycopg2==2.7.7
PyInstaller==3.4
Pypubsub==4.0.0
python-dateutil==2.7.3
python-docx==0.8.7
six==1.11.0
virtualenv==16.4.3
wxPython==4.0.3
xlrd==1.1.0
